backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
apscheduler
playwright
playwright-stealth
python-dotenv
